django/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==2.4.0
django-filter==2.4.0
django-ordered-model==3.4.3
django-sendgrid-v5==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
requests==2.23.0
requests-oauthlib==1.3.1
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.25.11